refactor: move action functions to video list composable
